requirements.txt
pytest
langchain
langchain-core
langchain-community
langchain-openai
langgraph

python-dotenv
pypdf
sentence_transformers
langchainhub
